+++ retail/projects/models.py
@@ -4,7 +4,7 @@ from django.core.cache import cache
 
 class Project(models.Model):
     name = models.CharField(max_length=256)
-    uuid = models.UUIDField()
+    uuid = models.UUIDField(unique=True)
     organization_uuid = models.UUIDField(null=True)
     vtex_account = models.CharField(max_length=100, null=True, blank=True)
     config = models.JSONField(default=dict)