+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+# 3.20.3
+
+## *Feature*
+  - Add beat entrypoint
+
+## *Fix*
+  - fix: remove template_variable from required fields
+
+# 3.20.2
+
+## *Fix*
+  - fix: unify template search by name and version_name in get_current_template
+
+# 3.20.1
+
+## *Fix*
+  - fix: adjust method to check integrated delivered order template
+
+# 3.20.0
+
+## *Feature*
+  - feat: implement delivered order tracking functionality
+
+# 3.19.2
+
+## *Feature*
+  - feat: add configuration to ignore empty carts for specific projects
+
+# 3.19.1
+
+## *Feature*
+  - feat: add logs to broadcast service
+
+# 3.19.0
+
+## *Feature*
+  - feat: ensure unique UUID for projects and improve message consumption logging
+  - feat: add parent_agent_uuid field to IntegratedAgent model and update order status use case logic
+  - feat: add Celery Beat schedule for cleaning up old carts
+  - refactor: fetch_templates_from_user
+
+## *Fix*
+  - fix: Use direct imports in template model
+
+# 3.18.6
+
+## *Feature*
+  - feat: setup elastic apm
+  - feat: clears cache after updating integrated agent
+  - feat: add cors config
+
 # 3.18.5
 
 ## *Feature*